# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: API E-commerce Detailing
  description: Documentación de la API del proyecto final de Backend
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
...
  /api/users:
    get:
      summary: Obtener todos los usuarios
      responses:
        "200":
          description: Lista de usuarios
    post:
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Usuario creado
...
  /api/products:
    get:
      summary: Obtener todos los productos
      responses:
        "200":
          description: Lista de productos
    post:
      summary: Crear un producto nuevo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Producto creado
...
  /api/carts/{cid}:
    get:
      summary: Obtener productos de un carrito
      parameters:
        - name: cid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Carrito encontrado
        "404":
          description: Carrito no encontrado
...
components:
  schemas:
    User:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [user, admin]
    Product:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        category:
          type: string
        thumbnails:
          type: array
          items:
            type: string
...